pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icedrive-authentication"
version = "0.0.1"
authors = [
  { name="Alejandro Cañas Borreguero", email="youremailaddress@uclm.es" },
]
description = "Implementation of IceDrive Authentication service"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-only"
dependencies = [
    "zeroc-ice>=3.7",
]

[project.optional-dependencies]
tests = ["pytest"]
linters = ["black", "ruff", "pydocstyle"]

[project.urls]
"Homepage" = "https://github.com/YOURUSERNAME/YOURREPOSITORY"

[project.scripts]
icedrive-authentication-server = "icedrive_authentication.command_lines_handlers:server"
icedrive-authentication-client = "icedrive_authentication.command_lines_handlers:client"

